+++ src/tty/mod.rs
@@ -59,3 +59,9 @@ pub fn draw(
     }
     Ok(())
 }
+pub fn csi(
+    fg: u8,
+    bg: u8,
+) -> String {
+    format!("\u{1b}[1m\u{1b}[38;5;{}m\u{1b}[48;5;{}m", fg, bg)
+}